Organize pacakges and setup tests
