+++ jestSetup.ts
@@ -1,0 +1,1 @@
+import './src/i18n';
