+++ package.json
@@ -1,38 +1,48 @@
 {
-	"name": "svelte-routify-typescript",
-	"private": true,
-	"scripts": {
-		"dev": "run-p dev:*",
-		"dev:routify": "routify",
-		"dev:vite": "vite",
-		"preview": "vite preview",
-		"build": "run-s build:*",
-		"build:routify": "routify -b",
-		"build:vite": "vite build"
-	},
-	"devDependencies": {
-		"@roxi/routify": "^2.18.3",
-		"@sveltejs/vite-plugin-svelte": "^1.0.0-next.5",
-		"@tsconfig/svelte": "^2.0.1",
-		"npm-run-all": "^4.1.5",
-		"postcss": "^8.4.4",
-		"postcss-cssnext": "^3.1.1",
-		"postcss-custom-media": "^8.0.0",
-		"postcss-import": "^14.0.2",
-		"postcss-load-config": "^3.1.0",
-		"postcss-nested": "^5.0.6",
-		"postcss-plugin": "^1.0.0",
-		"postcss-preset-env": "^7.0.1",
-		"precss": "^4.0.0",
-		"svelte": "^3.35.0",
-		"svelte-preprocess": "^4.7.0",
-		"typescript": "^4.5.2",
-		"vite": "^2.0.5"
-	},
-	"dependencies": {
-		"@dooboo/eslint-config-svelte": "^0.0.7",
-		"eslint": "^8.4.0",
-		"svelte-i18n": "^3.3.13",
-		"svelte-theme": "^1.4.3"
-	}
+  "name": "svelte-routify-typescript",
+  "private": true,
+  "scripts": {
+    "dev": "run-p dev:*",
+    "dev:routify": "routify",
+    "dev:vite": "vite",
+    "preview": "vite preview",
+    "build": "run-s build:*",
+    "build:routify": "routify -b",
+    "build:vite": "vite build",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "lint": "eslint src --ext .ts,.js,.svelte",
+    "format": "prettier --write ./src/**/*.{js,svelte,html,ts}",
+    "test": "jest"
+  },
+  "dependencies": {
+    "svelte-i18n": "^3.3.13",
+    "svelte-theme": "^1.4.3"
+  },
+  "devDependencies": {
+    "@roxi/routify": "^2.18.3",
+    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.5",
+    "@testing-library/svelte": "^3.0.3",
+    "@tsconfig/svelte": "^2.0.1",
+    "@types/jest": "^27.0.3",
+    "babel-jest": "^27.4.2",
+    "jest": "^27.4.3",
+    "postcss": "^8.4.4",
+    "postcss-cssnext": "^3.1.1",
+    "postcss-custom-media": "^8.0.0",
+    "postcss-import": "^14.0.2",
+    "postcss-load-config": "^3.1.0",
+    "postcss-nested": "^5.0.6",
+    "postcss-plugin": "^1.0.0",
+    "postcss-preset-env": "^7.0.1",
+    "precss": "^4.0.0",
+    "svelte": "^3.35.0",
+    "svelte-jester": "^2.1.5",
+    "svelte-preprocess": "^4.7.0",
+    "ts-jest": "^27.0.7",
+    "ts-node": "10.4.0",
+    "typescript": "^4.5.2",
+    "vite": "^2.0.5",
+    "@dooboo/eslint-config-svelte": "^0.0.7",
+    "eslint": "^8.4.0"
+  }
 }
