+++ src/layouts/__tests__/button.test.ts
@@ -1,0 +1,8 @@
+import Button from '../button.svelte';
+import {render} from '@testing-library/svelte';
+
+it('should render without error', () => {
+  const testngLib = render(Button);
+
+  expect(testngLib.container).toMatchSnapshot();
+});
