+++ tsconfig.json
@@ -3,7 +3,6 @@
   "include": [
     "src/**/*",
     "babel.config.js",
-    "jest.config.js",
     "svelte.config.js",
     "postcss.config.js",
     "jestSetup.ts",
@@ -12,6 +11,20 @@
   ],
   "exclude": ["node_modules/*", "dist"],
   "compilerOptions": {
-    "resolveJsonModule": true
+    "outDir": "build",
+    "resolveJsonModule": true,
+    "lib": ["DOM", "ES2017", "WebWorker"],
+    "esModuleInterop": true,
+    "importsNotUsedAsValues": "remove",
+    "strict": true,
+    "baseUrl": "./src",
+    "allowSyntheticDefaultImports": true,
+    "experimentalDecorators": true,
+    "forceConsistentCasingInFileNames": true,
+    "importHelpers": true,
+    "skipLibCheck": true,
+    "noEmitHelpers": true,
+    "noImplicitReturns": true,
+    "noUnusedLocals": false
   }
 }
